changes for calomentor w/matebot
