+++ src/interfaces/mentorship.interface.ts
@@ -1,6 +1,3 @@
-import { ITimeSlot } from './timeslot.interface';
-import { IWarning } from './warning.interface';
-
 export enum STATUS {
   ACTIVE = 'ACTIVE',
   CANCEL = 'CANCEL',
@@ -14,27 +11,28 @@ export enum WHOCANCELED {
 }
 
 export interface IMentorship {
-  id: string;
   mentee_email: string;
   mentee_username_discord: string;
   mentor_email: string;
   feedback_mentee: string;
   mentor_id: string;
   mentor_name: string;
+  mentor_username_discord?: string;
   info: string;
   mentee_name: string;
   mentorship_token: string;
   mentee_id: string;
   mentorship_status: STATUS;
+  id?: string;
   cancel_cause: string;
   who_canceled: WHOCANCELED;
-  time_slot_info?: ITimeSlot;
   time_slot_id?: string;
   feedback_stars: 1 | 2 | 3 | 4 | 5;
   feedback_mentee_private?: string;
-  warning_info?: IWarning;
   mentee_timezone: string;
   mentor_timezone?: string;
+  mentorship_create_date?: string;
+  from_bot: boolean
 }
 
 export interface ICancelForm {
